@@ -1,26 +1,26 @@
 ﻿using MauiPageFullScreen;
 using Microsoft.Extensions.Logging;
 
-namespace MauiFullScreenSample
+[assembly: XamlCompilation(XamlCompilationOptions.Compile)]
+namespace MauiFullScreenSample;
+
+public static class MauiProgram
 {
-    public static class MauiProgram
+    public static MauiApp CreateMauiApp()
     {
-        public static MauiApp CreateMauiApp()
-        {
-            var builder = MauiApp.CreateBuilder();
-            builder
-                .UseMauiApp<App>().UseFullScreen()
-                .ConfigureFonts(fonts =>
-                {
-                    fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
-                    fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
-                });
+        var builder = MauiApp.CreateBuilder();
+        builder
+            .UseMauiApp<App>().UseFullScreen()
+            .ConfigureFonts(fonts =>
+            {
+                fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
+                fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
+            });
 
 #if DEBUG
-            builder.Logging.AddDebug();
+        builder.Logging.AddDebug();
 #endif
-
-            return builder.Build();
-        }
+		builder.Services.AddSingleton<AppShell>();
+		return builder.Build();
     }
 }
