@@ -1,10 +1,9 @@
 ﻿using Foundation;
 
-namespace MauiFullScreenSample
-{
+namespace MauiFullScreenSample;
+
     [Register("AppDelegate")]
     public class AppDelegate : MauiUIApplicationDelegate
     {
         protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
     }
-}
