@@ -1,4 +1,6 @@
 ﻿
+using Android.Views;
+using AndroidX.Core.View;
 using MauiPageFullScreen.Interface;
 using Platform = Microsoft.Maui.ApplicationModel.Platform;
 
@@ -8,40 +10,30 @@ using Views = AndroidX.Core.View;
 
 namespace MauiPageFullScreen.Devices;
 
-internal class DeviceServices : IDeviceServices
+class DeviceServices : IDeviceServices
 {
+    static Android.Views.Window window => Platform.CurrentActivity?.Window ?? throw new InvalidOperationException("Current activity is null");
+    static Android.Views.View decorView => window.DecorView ?? throw new InvalidOperationException("DecorView is null");
+    static AndroidX.Core.View.WindowInsetsControllerCompat insetsController => WindowCompat.GetInsetsController(window, decorView) ?? throw new InvalidOperationException("InsetsController is null");
     public void RestoreScreen()
     {
-        PageExtensions.SetBarStatus(false);
-        var activity = Platform.CurrentActivity;
-
-        if (activity == null || activity.Window == null)
+        window.ClearFlags(WindowManagerFlags.Fullscreen);
+        window.SetFlags(WindowManagerFlags.DrawsSystemBarBackgrounds, WindowManagerFlags.DrawsSystemBarBackgrounds);
+        insetsController.SystemBarsBehavior = WindowInsetsControllerCompat.BehaviorDefault;
+        if (OperatingSystem.IsAndroidVersionAtLeast(34))
         {
-            return;
+            insetsController.Show(WindowInsets.Type.SystemBars());
         }
-
-        Views.WindowCompat.SetDecorFitsSystemWindows(activity.Window, false);
-        var windowInsetsControllerCompat = Views.WindowCompat.GetInsetsController(activity.Window, activity.Window.DecorView);
-        var types = Views.WindowInsetsCompat.Type.StatusBars() |
-                    Views.WindowInsetsCompat.Type.NavigationBars();
-        windowInsetsControllerCompat.Show(types);
     }
     public void FullScreen()
     {
-        PageExtensions.SetBarStatus(true);
-        var activity = Platform.CurrentActivity;
-
-        if (activity == null || activity.Window == null)
+        window.ClearFlags(WindowManagerFlags.LayoutNoLimits);
+        window.AddFlags(WindowManagerFlags.Fullscreen);
+        window.SetFlags(WindowManagerFlags.Fullscreen, WindowManagerFlags.Fullscreen);
+        insetsController.SystemBarsBehavior = WindowInsetsControllerCompat.BehaviorShowTransientBarsBySwipe;
+        if (OperatingSystem.IsAndroidVersionAtLeast(34))
         {
-            return;
+            insetsController.Hide(WindowInsets.Type.SystemBars());
         }
-
-        Views.WindowCompat.SetDecorFitsSystemWindows(activity.Window, false);
-        var windowInsetsControllerCompat = Views.WindowCompat.GetInsetsController(activity.Window, activity.Window.DecorView);
-        var types = Views.WindowInsetsCompat.Type.StatusBars() |
-                    Views.WindowInsetsCompat.Type.NavigationBars();
-
-        windowInsetsControllerCompat.SystemBarsBehavior = Views.WindowInsetsControllerCompat.BehaviorShowBarsBySwipe;
-        windowInsetsControllerCompat.Hide(types);
     }
 }
