@@ -8,10 +8,10 @@ namespace MauiPageFullScreen.Interface;
 interface IDeviceServices
 {
 #if WINDOWS
-    public AppWindow FullScreen();
-    public AppWindow RestoreScreen();
+    AppWindow FullScreen();
+    AppWindow RestoreScreen();
 #else
-    public void FullScreen();
-    public void RestoreScreen();
+    void FullScreen();
+    void RestoreScreen();
 #endif
 }
